Featire-IngresarReceta
-Se cambia el Nombre de RecetaFrame por IngrsarReceta

-Se agrega opcion eliminar filas

-Se agrega opcion guardar receta a la base de datos
